stamp_pure_proj/requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
matplotlib==3.5.2
numpy==1.23.0
opencv-python==4.6.0.66
Pillow==9.1.1
PyYAML==6.0
requests==2.28.1
scipy==1.8.1
torch==1.7.1
torchvision==0.8.2
tqdm==4.64.0

# Logging -------------------------------------
tensorboard==2.9.1
# wandb

# Plotting ------------------------------------
pandas==1.4.3
seaborn==0.11.2


# others
scikit-learn==0.24.2
xlrd==1.2.0
joblib==1.1.0
xgboost==1.6.1
lightgbm==3.3.2
PyQt5==5.14.2
# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export

# Extras --------------------------------------
# albumentations>=1.0.3
# Cython  # for pycocotools https://github.com/cocodataset/cocoapi/issues/172
# pycocotools>=2.0  # COCO mAP
# roboflow
thop  # FLOPs computation
